Portfolio Insights Generator/requirements.txt
python-dotenv
openai>=1.0
streamlit>=1.30
pandas
